+++ src/test/java/io/jstuff/log/test/LoggerTest.java
@@ -2,7 +2,7 @@
  * @(#) LoggerTest.java
  *
  * log-front-api  Logging Interface API
- * Copyright (c) 2022 Peter Wall
+ * Copyright (c) 2022, 2025 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -27,6 +27,7 @@ package io.jstuff.log.test;
 
 import java.time.Clock;
 import java.time.Instant;
+import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 
 import org.junit.Test;
@@ -50,10 +51,48 @@ public class LoggerTest {
     }
 
     @Test
+    public void shouldLogTrace() {
+        MockLogger mockLogger = new MockLogger("Lorikeet", Level.TRACE, LoggerFactory.systemClock);
+        mockLogger.trace("hello");
+        assertEquals("Lorikeet TRACE hello\n", mockLogger.getContents());
+    }
+
+    @Test
+    public void shouldLogTraceWithTime() {
+        MockLogger mockLogger = new MockLogger("Lorikeet", Level.TRACE, LoggerFactory.systemClock);
+        OffsetDateTime time = OffsetDateTime.of(2025, 7, 18, 12, 20, 24, 123000000, ZoneOffset.ofHours(10));
+        mockLogger.trace(time.toInstant(), "hello");
+        assertEquals("Lorikeet TRACE@2025-07-18T12:20:24.123+10:00 hello\n", mockLogger.getContents());
+    }
+
+    @Test
+    public void shouldLogDebug() {
+        MockLogger mockLogger = new MockLogger("Lorikeet", Level.DEBUG, LoggerFactory.systemClock);
+        mockLogger.debug("hello");
+        assertEquals("Lorikeet DEBUG hello\n", mockLogger.getContents());
+    }
+
+    @Test
+    public void shouldLogDebugWithTime() {
+        MockLogger mockLogger = new MockLogger("Lorikeet", Level.DEBUG, LoggerFactory.systemClock);
+        OffsetDateTime time = OffsetDateTime.of(2025, 7, 18, 12, 20, 24, 123000000, ZoneOffset.ofHours(10));
+        mockLogger.debug(time.toInstant(), "hello");
+        assertEquals("Lorikeet DEBUG@2025-07-18T12:20:24.123+10:00 hello\n", mockLogger.getContents());
+    }
+
+    @Test
     public void shouldLogInfo() {
         MockLogger mockLogger = new MockLogger("Galah", Level.INFO, LoggerFactory.systemClock);
         mockLogger.info("hello");
         assertEquals("Galah INFO hello\n", mockLogger.getContents());
+    }
+
+    @Test
+    public void shouldLogInfoWithTime() {
+        MockLogger mockLogger = new MockLogger("Galah", Level.INFO, LoggerFactory.systemClock);
+        OffsetDateTime time = OffsetDateTime.of(2025, 7, 18, 12, 20, 24, 123000000, ZoneOffset.ofHours(10));
+        mockLogger.info(time.toInstant(), "hello");
+        assertEquals("Galah INFO@2025-07-18T12:20:24.123+10:00 hello\n", mockLogger.getContents());
     }
 
     @Test
@@ -64,13 +103,36 @@ public class LoggerTest {
     }
 
     @Test
+    public void shouldLogInfoWithTimeUsingLambda() {
+        MockLogger mockLogger = new MockLogger("Corella", Level.INFO, LoggerFactory.systemClock);
+        OffsetDateTime time = OffsetDateTime.of(2025, 7, 18, 12, 20, 24, 123000000, ZoneOffset.ofHours(10));
+        mockLogger.info(time.toInstant(), () -> "hello");
+        assertEquals("Corella INFO@2025-07-18T12:20:24.123+10:00 hello\n", mockLogger.getContents());
+    }
+
+    @Test
     public void shouldNotLogInfoWhenLevelSetToWarn() {
         MockLogger mockLogger = new MockLogger("Cockatoo", Level.WARN, LoggerFactory.systemClock);
         mockLogger.info(() -> {
             fail("Should not be called");
             return null;
         });
-        assertEquals("", mockLogger.getContents());
+        assertEquals(0, mockLogger.getContents().length());
+    }
+
+    @Test
+    public void shouldLogWarn() {
+        MockLogger mockLogger = new MockLogger("Galah", Level.INFO, LoggerFactory.systemClock);
+        mockLogger.warn("hello");
+        assertEquals("Galah WARN hello\n", mockLogger.getContents());
+    }
+
+    @Test
+    public void shouldLogWarnWithTime() {
+        MockLogger mockLogger = new MockLogger("Galah", Level.INFO, LoggerFactory.systemClock);
+        OffsetDateTime time = OffsetDateTime.of(2025, 7, 18, 12, 20, 24, 123000000, ZoneOffset.ofHours(10));
+        mockLogger.warn(time.toInstant(), "hello");
+        assertEquals("Galah WARN@2025-07-18T12:20:24.123+10:00 hello\n", mockLogger.getContents());
     }
 
     @Test
@@ -79,6 +141,15 @@ public class LoggerTest {
         Throwable throwable = new RuntimeException("magic");
         mockLogger.error(throwable, () -> "goodbye");
         assertEquals("Budgerigar ERROR goodbye : magic\n", mockLogger.getContents());
+    }
+
+    @Test
+    public void shouldLogErrorWithTimeAndThrowable() {
+        MockLogger mockLogger = new MockLogger("Budgerigar", Level.INFO, LoggerFactory.systemClock);
+        OffsetDateTime time = OffsetDateTime.of(2025, 7, 18, 12, 20, 24, 123000000, ZoneOffset.ofHours(10));
+        Throwable throwable = new RuntimeException("magic");
+        mockLogger.error(time.toInstant(), throwable, () -> "goodbye");
+        assertEquals("Budgerigar ERROR@2025-07-18T12:20:24.123+10:00 goodbye : magic\n", mockLogger.getContents());
     }
 
     @Test